// types/app.ts
export interface AppInfo {
  title: string
  description: string
  default_language: Locale
  copyright?: string
  privacy_policy?: string
  disable_session_same_site?: boolean
  robot_avatar?: string
  logo?: string
}